Add support for cached file to curl
This is to allow a resume after SWUpdate was interrupted and it is
restarted, for example after a power-cut. If the SWU is also saved with
the --output parameter, it should be possible to download just the rest
of data instead of restaring from the beginning.

The data in the cached file will be read first, and then a RESUME to the
server is sent to go on with the rest of data.

Signed-off-by: Stefano Babic <sbabic@denx.de>